Task 14 - Fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-unsigned long long int n1 = 0, n2 = 1, sum;
+unsigned long n1 = 0, n2 = 1, sum;
 int count = 0;
 
 while (count <= 50)
@@ -17,11 +17,11 @@ n1 = n2;
 n2 = sum;
 if (count == 50)
 {
-printf("%lld", sum);
+printf("%ld", sum);
 }
 else
 {
-printf("%lld, ", sum);
+printf("%ld, ", sum);
 }
 count++;
 }
